+++ Assets/DeltaDNA/DDNA.cs
@@ -428,14 +428,38 @@ namespace DeltaDNA
             var notificationEvent = new GameEvent("notificationOpened");
             try {
                 if (payload.ContainsKey("_ddId")) {
-                    notificationEvent.AddParam("notificationId", Convert.ToInt32(payload["_ddId"]));
+                    notificationEvent.AddParam("notificationId", Convert.ToInt64(payload["_ddId"]));
                 }
                 if (payload.ContainsKey("_ddName")) {
                     notificationEvent.AddParam("notificationName", payload["_ddName"]);
                 }
+
+                bool insertCommunicationAttrs = false;
+                if (payload.ContainsKey("_ddCampaign")) {
+                    notificationEvent.AddParam("campaignId", Convert.ToInt64(payload["_ddCampaign"]));
+                    insertCommunicationAttrs = true;
+                }
+                if (payload.ContainsKey("_ddCohort")) {
+                    notificationEvent.AddParam("cohortId", Convert.ToInt64(payload["_ddCohort"]));
+                    insertCommunicationAttrs = true;
+                }
+                if (insertCommunicationAttrs && payload.ContainsKey("_ddCommunicationSender")) {
+                    // _ddCommunicationSender inserted by respective native notification SDK
+                    notificationEvent.AddParam("communicationSender", payload["_ddCommunicationSender"]);
+                    notificationEvent.AddParam("communicationState", "OPEN");
+                }
+
                 if (payload.ContainsKey("_ddLaunch")) {
+                    // _ddLaunch inserted by respective native notification SDK
                     notificationEvent.AddParam("notificationLaunch", Convert.ToBoolean(payload["_ddLaunch"]));
                 }
+                if (payload.ContainsKey("_ddCampaign")) {
+                    notificationEvent.AddParam("campaignId", Convert.ToInt64(payload["_ddCampaign"]));
+                }
+                if (payload.ContainsKey("_ddCohort")) {
+                    notificationEvent.AddParam("cohortId", Convert.ToInt64(payload["_ddCohort"]));
+                }
+                notificationEvent.AddParam("communicationState", "OPEN");
             } catch (Exception ex) {
                 Logger.LogError("Error parsing push notification payload. "+ex.Message);
             }
@@ -493,10 +517,17 @@ namespace DeltaDNA
         public AndroidNotifications AndroidNotifications { get; private set; }
 
         /// <summary>
-        /// Clears the persistent data such as user id.  Useful for testing purposes.
+        /// Clears the persistent data, such as user id. The SDK should be stopped
+        /// before this method is called.
+        /// 
+        /// Useful for testing purposes.
         /// </summary>
         public void ClearPersistentData()
         {
+            if (HasStarted) {
+                Logger.LogWarning("SDK has not been stopped before clearing persistent data");
+            }
+
             PlayerPrefs.DeleteKey(PF_KEY_USER_ID);
 
             if (this.eventStore != null) {